requirements.txt
# AI Hiring System - MVP Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP & Scraping
httpx==0.26.0
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# AI
openai==1.10.0

# Email
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0
